Fix dashboard-backend integration issues for dev mode
- Add NEXT_PUBLIC_GRAPHQL_URL to .env.development so the dashboard
  reaches the Go backend (port 8080) instead of the Next.js dev server
- Add GraphQL GET transport to gqlgen so query-string requests succeed
- Fix timezone-offset type mismatch (Flink returns a number, not string)
- Fix VertexMetrics type mismatch (Flink returns floats like 68.0, not
  integers) by switching int64 fields to float64 with an f64() helper
- Make output:"export" conditional on production builds so dynamic routes
  like /jobs/[id] work in dev mode without generateStaticParams
- Improve job detail error UI: vertically centered layout with icon badge
  and clearer visual hierarchy

+++ server/internal/flink/types.go
@@ -79,18 +79,20 @@ type JobPlan struct {
 }
 
 // VertexMetrics represents I/O metrics within a vertex entry.
+// Flink REST returns all numeric metric values as JSON floats (e.g. 68.0),
+// so we use float64 to avoid unmarshal errors.
 type VertexMetrics struct {
-	ReadBytes                int64 `json:"read-bytes"`
-	ReadBytesComplete        bool  `json:"read-bytes-complete"`
-	WriteBytes               int64 `json:"write-bytes"`
-	WriteBytesComplete       bool  `json:"write-bytes-complete"`
-	ReadRecords              int64 `json:"read-records"`
-	ReadRecordsComplete      bool  `json:"read-records-complete"`
-	WriteRecords             int64 `json:"write-records"`
-	WriteRecordsComplete     bool  `json:"write-records-complete"`
-	AccumulatedBackpressured int64 `json:"accumulated-backpressured-time"`
-	AccumulatedIdle          int64 `json:"accumulated-idle-time"`
-	AccumulatedBusy          int64 `json:"accumulated-busy-time"`
+	ReadBytes                float64 `json:"read-bytes"`
+	ReadBytesComplete        bool    `json:"read-bytes-complete"`
+	WriteBytes               float64 `json:"write-bytes"`
+	WriteBytesComplete       bool    `json:"write-bytes-complete"`
+	ReadRecords              float64 `json:"read-records"`
+	ReadRecordsComplete      bool    `json:"read-records-complete"`
+	WriteRecords             float64 `json:"write-records"`
+	WriteRecordsComplete     bool    `json:"write-records-complete"`
+	AccumulatedBackpressured float64 `json:"accumulated-backpressured-time"`
+	AccumulatedIdle          float64 `json:"accumulated-idle-time"`
+	AccumulatedBusy          float64 `json:"accumulated-busy-time"`
 }
 
 // Vertex represents a single vertex entry within GET /jobs/:jobid → vertices[].
